+++ packages/mantine/src/components/index.ts
@@ -7,7 +7,6 @@ export * from './header';
 export * from './inline-confirm';
 export * from './table';
 export * from './prompt';
-export * from './modal';
 export * from './modal-wizard';
 export * from './button';
 export * from './menu';