+++ packages/mantine/src/index.ts
@@ -2,6 +2,7 @@ import {Tuple} from '@mantine/core';
 
 import {PlasmaColors} from './theme/PlasmaColors';
 
+export * from '@mantine/notifications';
 export * from '@mantine/carousel';
 export * from '@mantine/core';
 export type {FormValidateInput} from '@mantine/form/lib/types';
@@ -9,16 +10,16 @@ export * from '@mantine/hooks';
 export * from '@tanstack/table-core';
 export * from './components';
 export * from '@mantine/form';
+export {type NotificationProps} from '@mantine/notifications';
 export {Pagination} from '@mantine/core';
 // explicitly overriding mantine components
 export {
     Header,
+    type HeaderProps,
     Table,
     type TableProps,
     type TableState,
     type InitialTableState,
-    type HeaderProps,
-    Modal,
     Button,
     type ButtonProps,
     Menu,